LTS Ammo Update
Ammunition Production Customisation
- This tweak has been removed and replaced by Expensive Ammo.
Note: With the changes to ammo recipes in 1.6 I decided it was too troublesome and fiddly to leave it openly customisable.  So I decided to remake it from scratch with the goal of having 60 units of ammo created in each batch and having a Ingredient to Ammo ratio of 1.5.  No more fussing around with numbers just enable the tweak and go.

Expensive Ammo
- Changes the ammo crafting recipes from LTS to make ammunition more expensive to create.  The values are balanced around the idea that each ingredient will yield 1.5 units of ammo and each recipe will create 60 units of ammo.  The market price is automatically calculated based on the material costs.  Spacer ammo is made with a 1:2 ratio to make it somewhat cheaper than Ultratech ammo.